requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
twilio==6.37.0
